handle Analyze 7.5 files with > 32767 intensities
The official Analyze 7.5 format specification uses an int16 for the number of
intensity values (ni in the hdr file) and is therefore limited to 32767 values.
In IMS there are sometimes more than 32767 intensity values per spectrum. We use
the file size of the t2m file to find the correct numbers of intensity/mass
values and ignore the ni entry completely.
Thanks to Ken Frankel for reporting this problem.